requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
boto3
langchain
langchain-aws
langfuse==2.60.10
colornames
